+++ packages/internal-test-helpers/lib/confirm-export.ts
@@ -21,43 +21,42 @@ export default function confirmExport(
   exportName: string | { value: unknown; get: string; set: string },
   mod: any
 ) {
-  try {
-    let desc: PropertyDescriptor | null | undefined;
+  let desc: PropertyDescriptor | null | undefined;
 
-    if (path !== null) {
-      desc = getDescriptor(Ember, path);
-      assert.ok(desc, `the ${path} property exists on the Ember global`);
-    } else {
-      desc = null;
+  if (path !== null) {
+    desc = getDescriptor(Ember, path);
+    assert.ok(desc, `the ${path} property exists on the Ember global`);
+  } else {
+    desc = null;
+  }
+
+  if (desc == null) {
+    assert.notEqual(
+      mod[exportName as string],
+      undefined,
+      `${moduleId}#${exportName} is not \`undefined\``
+    );
+  } else if (typeof exportName === 'string') {
+    let value = 'value' in desc ? desc.value : desc.get!.call(Ember);
+    // Access early so we still trigger deprecation
+    let actual = mod[exportName];
+    // mod[exportName] has no name, but value has a name here.
+    // This could have something to do with how the proxy is working,
+    // but it's not all that important, as deprecate users will still have
+    // the same behavior, including the deprecation message
+    if (exportName !== 'deprecate') {
+      assert.equal(value, actual, `Ember.${path} is exported correctly`);
     }
+    assert.notEqual(mod[exportName], undefined, `Ember.${path} is not \`undefined\``);
+  } else if ('value' in desc) {
+    assert.equal(desc.value, exportName.value, `Ember.${path} is exported correctly`);
+  } else {
+    assert.equal(desc.get, mod[exportName.get], `Ember.${path} getter is exported correctly`);
+    assert.notEqual(desc.get, undefined, `Ember.${path} getter is not undefined`);
 
-    if (desc == null) {
-      assert.notEqual(
-        mod[exportName as string],
-        undefined,
-        `${moduleId}#${exportName} is not \`undefined\``
-      );
-    } else if (typeof exportName === 'string') {
-      let value = 'value' in desc ? desc.value : desc.get!.call(Ember);
-      assert.equal(value, mod[exportName], `Ember.${path} is exported correctly`);
-      assert.notEqual(mod[exportName], undefined, `Ember.${path} is not \`undefined\``);
-    } else if ('value' in desc) {
-      assert.equal(desc.value, exportName.value, `Ember.${path} is exported correctly`);
-    } else {
-      assert.equal(desc.get, mod[exportName.get], `Ember.${path} getter is exported correctly`);
-      assert.notEqual(desc.get, undefined, `Ember.${path} getter is not undefined`);
-
-      if (exportName.set) {
-        assert.equal(desc.set, mod[exportName.set], `Ember.${path} setter is exported correctly`);
-        assert.notEqual(desc.set, undefined, `Ember.${path} setter is not undefined`);
-      }
+    if (exportName.set) {
+      assert.equal(desc.set, mod[exportName.set], `Ember.${path} setter is exported correctly`);
+      assert.notEqual(desc.set, undefined, `Ember.${path} setter is not undefined`);
     }
-  } catch (error) {
-    assert.pushResult({
-      result: false,
-      message: `An error occurred while testing ${path} is exported from ${moduleId}`,
-      actual: error,
-      expected: undefined,
-    });
   }
 }
